requirements.txt
scanpy>=1.9
torch>=1.12
numpy>=1.21
pandas>=1.4
matplotlib>=3.5
seaborn>=0.12
scipy>=1.9
scikit-learn>=1.1
anndata>=0.8
tqdm
gprofiler-official>=1.0
